move userReacji state to chat2
