+++ shared/actions/chat2-gen.tsx
@@ -138,6 +138,7 @@ export const updateNotificationSettings = 'chat2:updateNotificationSettings'
 export const updateReactions = 'chat2:updateReactions'
 export const updateTeamRetentionPolicy = 'chat2:updateTeamRetentionPolicy'
 export const updateUnreadline = 'chat2:updateUnreadline'
+export const updateUserReacjis = 'chat2:updateUserReacjis'
 
 // Payload Types
 type _AddAttachmentViewMessagePayload = {
@@ -604,6 +605,7 @@ type _UpdateUnreadlinePayload = {
   readonly conversationIDKey: Types.ConversationIDKey
   readonly messageID: Types.MessageID
 }
+type _UpdateUserReacjisPayload = {readonly userReacjis: RPCTypes.UserReacjis | null}
 
 // Action Creators
 /**
@@ -1243,6 +1245,10 @@ export const createUpdateMoreToLoad = (payload: _UpdateMoreToLoadPayload): Updat
 export const createUpdateNotificationSettings = (
   payload: _UpdateNotificationSettingsPayload
 ): UpdateNotificationSettingsPayload => ({payload, type: updateNotificationSettings})
+export const createUpdateUserReacjis = (payload: _UpdateUserReacjisPayload): UpdateUserReacjisPayload => ({
+  payload,
+  type: updateUserReacjis,
+})
 
 // Action Payloads
 export type AddAttachmentViewMessagePayload = {
@@ -1719,6 +1725,10 @@ export type UpdateUnreadlinePayload = {
   readonly payload: _UpdateUnreadlinePayload
   readonly type: typeof updateUnreadline
 }
+export type UpdateUserReacjisPayload = {
+  readonly payload: _UpdateUserReacjisPayload
+  readonly type: typeof updateUserReacjis
+}
 
 // All Actions
 // prettier-ignore
@@ -1853,4 +1863,5 @@ export type Actions =
   | UpdateReactionsPayload
   | UpdateTeamRetentionPolicyPayload
   | UpdateUnreadlinePayload
+  | UpdateUserReacjisPayload
   | {type: 'common:resetStore', payload: null}
