+++ shared/chat/conversation/messages/react-button/emoji-picker/container.tsx
@@ -28,7 +28,7 @@ type WrapperState = {
 }
 
 const mapStateToProps = (state, ownProps: OwnProps) => ({
-  topReacjis: state.config.userReacjis.topReacjis,
+  topReacjis: state.chat2.userReacjis.topReacjis,
 })
 
 const mapDispatchToProps = (dispatch, ownProps: OwnProps) => {
