+++ shared/constants/chat2/index.tsx
@@ -18,6 +18,9 @@ import {formatTextForQuoting} from '../../util/chat'
 import * as Router2 from '../router2'
 import HiddenString from '../../util/hidden-string'
 
+export const defaultTopReacjis = [':+1:', ':-1:', ':tada:', ':joy:', ':sunglasses:']
+const defaultSkinTone = 1
+
 export const makeState = I.Record<Types._State>({
   accountsInfoMap: I.Map(),
   attachmentFullscreenSelection: null,
@@ -58,6 +61,7 @@ export const makeState = I.Record<Types._State>({
   unfurlPromptMap: I.Map(),
   unreadMap: I.Map(),
   unsentTextMap: I.Map(),
+  userReacjis: {skinTone: defaultSkinTone, topReacjis: defaultTopReacjis},
 
   // Team Building
   ...TeamBuildingConstants.makeSubState(),
