+++ shared/constants/types/config.tsx
@@ -1,5 +1,4 @@
 import * as I from 'immutable'
-import * as RPCTypes from '../../constants/types/rpc-gen'
 import {ConversationIDKey} from './chat2'
 import {Tab} from '../tabs'
 import {DeviceID} from './rpc-gen'
@@ -57,6 +56,5 @@ export type _State = {
   uid: string
   userActive: boolean
   username: string
-  userReacjis: RPCTypes.UserReacjis
 }
 export type State = I.RecordOf<_State>
